chore: combine text input and area to one component
