--- src/components/forms/text-area/text-area.stories.mdx
@@ -1,23 +1,0 @@
-import { ArgsTable, Canvas, Meta, Story } from '@storybook/addon-docs';
-import { TextArea } from './text-area';
-
-<Meta
-  title="Components/Forms/TextArea"
-  component={{ TextArea }} />
-
-# TextArea
-A styled version of the HTML `textarea` with some extra functionalities. Mainly used in the contact and career-details page.
-
-<Canvas withSource='open'>
-  <Story name='Regular'>
-    <TextArea
-      label="Message"
-      name="message"
-      errorMessage="Error Message"
-      required={true}
-    />
-  </Story>
-</Canvas>
-
-<ArgsTable of={TextArea} />
-
