+++ src/components/forms/textinput/textinput.tsx
@@ -3,7 +3,40 @@ import styled from 'styled-components';
 import { TextStyles } from '../../typography';
 import { theme } from '../../layout/theme';
 
-export const Label = styled.label`
+const TextArea = styled.textarea<{ hasError?: boolean }>`
+  height: 160px;
+  width: 100%;
+  padding: 19px 16px;
+  resize: vertical;
+
+  ${TextStyles.textR}
+
+  background: ${theme.palette.background.leadbox};
+  border: 2px solid
+    ${(props) =>
+      props.hasError
+        ? theme.palette.text.errorMessage
+        : theme.palette.background.leadbox};
+`;
+
+const Input = styled.input<{ hasError?: boolean }>`
+  width: 320px;
+  height: 48px;
+
+  padding-left: 16px;
+  padding-right: 16px;
+
+  ${TextStyles.textR}
+
+  background: ${theme.palette.background.leadbox};
+  border: 2px solid
+    ${(props) =>
+      props.hasError
+        ? theme.palette.text.errorMessage
+        : theme.palette.background.leadbox};
+`;
+
+const Label = styled.label`
   ${TextStyles.textXS}
   display: block;
   margin-bottom: 4px;
@@ -11,7 +44,7 @@ export const Label = styled.label`
   color: ${theme.palette.text.default};
 `;
 
-export const StyledErrorMessage = styled.span`
+const StyledErrorMessage = styled.span`
   ${TextStyles.textXS}
   font-weight: 700;
   display: block;
@@ -20,35 +53,63 @@ export const StyledErrorMessage = styled.span`
   color: ${theme.palette.text.errorMessage};
 `;
 
-interface TextInputProps {
-  errorMessage: string;
-  label: string;
+export interface TextAreaProps {
+  inputRef: any;
+  /**
+   * `name` is the id of the text input
+   */
+  name: string;
+  /**
+   * A label which is displayed above the text input
+   */
+  label?: string;
+  /**
+   * Render the TextInput component as HTML `input` or `textarea`
+   */
+  type?: 'textarea' | 'input';
+  /**
+   * The error message is displayed below the text area and changes the style of the text input
+   */
+  errorMessage?: string;
+  /**
+   * Marks the component as a required field of the form
+   */
+  required?: boolean;
 }
 
-export const TextInput = ({ label, errorMessage }: TextInputProps) => {
-  const StyledInputText = styled.input.attrs({
-    type: 'input',
-  })`
-    width: 320px;
-    height: 48px;
-
-    padding-left: 16px;
-    padding-right: 16px;
-
-    ${TextStyles.textR}
-    letter-spacing: -0.01em;
-    color: #202840;
-
-    border: 2px solid;
-    border-color: ${errorMessage ? '#FF0D35' : '#f7f8fa'};
-    background: #f7f8fa;
-  `;
-
+export const TextInput = ({
+  label,
+  name,
+  errorMessage,
+  inputRef,
+  required,
+  type = 'input',
+}: TextAreaProps) => {
   return (
-    <label>
-      <Label>{label}</Label>
-      <StyledInputText />
+    <>
+      {label && (
+        <Label htmlFor={name}>
+          {label} {required && <span aria-hidden={true}>*</span>}
+        </Label>
+      )}
+      {type === 'textarea' ? (
+        <TextArea
+          {...inputRef}
+          hasError={errorMessage}
+          id={name}
+          aria-required={required}
+        />
+      ) : (
+        <Input
+          type={'text'}
+          {...inputRef}
+          hasError={errorMessage}
+          id={name}
+          aria-required={required}
+        />
+      )}
+
       {errorMessage && <StyledErrorMessage>{errorMessage}</StyledErrorMessage>}
-    </label>
+    </>
   );
 };
